Added the "opt" series of functions
These functions each wrap one of the standard "user" functions,
and are set up expecting to be used inside a program/function that
*may* receive one or more inputs, and if not, should prompt the
user for it interactively.

Also modified user_onoff_list: it's default behavior now is to
return a list of the subset of options selected, this can be
switched back to the old way of returning a list of booleans
with the returntype keyword.